Subindo alterações no banco, para inplementação de finção cadastro

+++ cosmos/Cosmos-API/src/main/java/com/planbtech/cosmos/model/Interfaces/IAbilitie.java
@@ -11,14 +11,14 @@ public interface IAbilitie {
      *
      * @return {@code int }o codigo da habilidade
      */
-    public int getCode();
+    public Long getCode();
 
     /**
      * Metodo para criar ou alterar o codigo da habilidade
      *
      * @param code recebe o codigo a ser inserido
      */
-    public void setCode(int code);
+    public void setCode(Long code);
 
     /**
      * Metodo para pegar o nome da habilidade
